Improvements on notifications

+++ src/commonMain/kotlin/chat/sphinx/concepts/notification/SphinxNotificationManager.kt
@@ -22,7 +22,7 @@ interface SphinxNotificationManager {
         message: String
     )
 
-    fun clearNotification(notificationId: Int)
+    fun clearNotification(notificationId: Long)
 
     suspend fun toast(
         windowTitle: String,
